[Support #1447] Harmonize public term detail view initial language resolution with private view.
Also minor refactoring in term languages resolution.
