+++ src/component/multilingual/LanguageSelector.tsx
@@ -1,10 +1,9 @@
 import * as React from "react";
 import ISO6391 from "iso-639-1";
 import withI18n, {HasI18n} from "../hoc/withI18n";
-import Term, {TERM_MULTILINGUAL_ATTRIBUTES, TermData} from "../../model/Term";
+import Term from "../../model/Term";
 import {injectIntl} from "react-intl";
 import {Nav, NavItem, NavLink} from "reactstrap";
-import Utils from "../../util/Utils";
 import "./LanguageSelector.scss";
 import {FaTimesCircle} from "react-icons/fa";
 
@@ -12,16 +11,6 @@ interface LanguageSelectorProps extends HasI18n {
     term: Term | null;
     language: string;
     onSelect: (lang: string) => void;
-}
-
-export function getLanguages(term: Term | TermData): string[] {
-    const languages: Set<string> = new Set();
-    TERM_MULTILINGUAL_ATTRIBUTES.filter(att => term[att]).forEach(att => {
-        Utils.sanitizeArray(term[att]).forEach(attValue => Object.getOwnPropertyNames(attValue).forEach(n => languages.add(n)))
-    });
-    const langArr = Array.from(languages);
-    langArr.sort();
-    return langArr;
 }
 
 export function renderLanguages(languages: string[], selectedLanguage: string, formatMessage: (msgId: string, values: {} | undefined) => string, onSelect: (lang: string) => void, onRemove?: (lang: string) => void) {
@@ -49,7 +38,7 @@ const LanguageSelector: React.FC<LanguageSelectorProps> = props => {
     if (!term) {
         return null;
     }
-    const languages = getLanguages(term);
+    const languages = Term.getLanguages(term);
     if (languages.length <= 1) {
         return null;
     }
